Changed currency back to int - decimal didn't work

+++ db/yarn_store.sql
@@ -7,7 +7,7 @@ CREATE TABLE manufacturers (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255),
     last_payment_date DATE,
-    balance_due DECIMAL
+    balance_due INT
 );
 
 CREATE TABLE yarns (
@@ -19,8 +19,8 @@ CREATE TABLE yarns (
     length_metres INT,
     needle_size_mm DECIMAL,
     fibre_type VARCHAR(255),
-    buy_cost DECIMAL,
-    sell_price DECIMAL,
+    buy_cost INT,
+    sell_price INT,
     image VARCHAR(255)
 );
